magic mv to shared dir
