+++ webui/react/src/components/Router.tsx
@@ -2,11 +2,12 @@ import React, { ReactNode, useEffect, useState } from 'react';
 import { Redirect, Switch } from 'react-router-dom';
 import { CompatRoute } from 'react-router-dom-v5-compat';
 
-import { StoreActionUI, useStore, useStoreDispatch } from 'contexts/Store';
+import { useStore, useStoreDispatch } from 'contexts/Store';
 import useAuthCheck from 'hooks/useAuthCheck';
 import { paths } from 'routes/utils';
 import { RouteConfig } from 'shared/types';
 import { filterOutLoginLocation } from 'shared/utils/routes';
+import { StoreActionUI } from '../shared/contexts/UIStore';
 
 interface Props {
   routes: RouteConfig[];
