+++ webui/react/src/contexts/Store.tsx
@@ -1,16 +1,24 @@
-import React, { Dispatch, useContext, useReducer } from 'react';
-
 import { globalStorage } from 'globalStorage';
+import React, { Dispatch, useContext, useReducer } from 'react';
 import { V1UserWebSetting } from 'services/api-ts-sdk';
-import { DarkLight, Mode, Theme } from 'shared/themes';
 import { clone, isEqual } from 'shared/utils/data';
 import rootLogger from 'shared/utils/Logger';
 import { percent } from 'shared/utils/number';
 import {
-  Agent, Auth, ClusterOverview, ClusterOverviewResource,
-  DetailedUser, DeterminedInfo, PoolOverview, ResourcePool, ResourceType,
-  UserAssignment, UserRole, Workspace,
+  Agent,
+  Auth,
+  ClusterOverview,
+  ClusterOverviewResource,
+  DetailedUser,
+  DeterminedInfo,
+  PoolOverview,
+  ResourcePool,
+  ResourceType,
+  UserAssignment,
+  UserRole,
+  Workspace,
 } from 'types';
+import { ActionUI, initUI, reducerUI, StateUI } from '../shared/contexts/UIStore';
 
 const logger = rootLogger.extend('store');
 
@@ -21,79 +29,6 @@ interface Props {
 interface OmnibarState {
   isShowing: boolean;
 }
-
-interface StateUI {
-  chromeCollapsed: boolean;
-  darkLight: DarkLight;
-  isPageHidden: boolean;
-  mode: Mode;
-  showChrome: boolean;
-  showSpinner: boolean;
-  theme: Theme;
-}
-
-const initUI: StateUI = {
-  chromeCollapsed: false,
-  darkLight: DarkLight.Light,
-  isPageHidden: false,
-  mode: Mode.System,
-  showChrome: true,
-  showSpinner: false,
-  theme: {} as Theme,
-};
-
-export enum StoreActionUI {
-  HideUIChrome = 'HideUIChrome',
-  HideUISpinner = 'HideUISpinner',
-  SetMode = 'SetMode',
-  SetPageVisibility = 'SetPageVisibility',
-  SetTheme = 'SetTheme',
-  ShowUIChrome = 'ShowUIChrome',
-  ShowUISpinner = 'ShowUISpinner',
-}
-
-type ActionUI =
-| { type: StoreActionUI.HideUIChrome }
-| { type: StoreActionUI.HideUISpinner }
-| { type: StoreActionUI.SetMode; value: Mode }
-| { type: StoreActionUI.SetPageVisibility; value: boolean }
-| { type: StoreActionUI.SetTheme; value: { darkLight: DarkLight, theme: Theme } }
-| { type: StoreActionUI.ShowUIChrome }
-| { type: StoreActionUI.ShowUISpinner }
-
-/**
- * return a part of the input state that should be updated.
- * @param state ui state
- * @param action
- * @returns
- */
-const reducerUI = (state: StateUI, action: ActionUI): Partial<StateUI> | void => {
-  switch (action.type) {
-    case StoreActionUI.HideUIChrome:
-      if (!state.showChrome) return;
-      return { showChrome: false };
-    case StoreActionUI.HideUISpinner:
-      if (!state.showSpinner) return;
-      return { showSpinner: false };
-    case StoreActionUI.SetMode:
-      return { mode: action.value };
-    case StoreActionUI.SetPageVisibility:
-      return { isPageHidden: action.value };
-    case StoreActionUI.SetTheme:
-      return {
-        darkLight: action.value.darkLight,
-        theme: action.value.theme,
-      };
-    case StoreActionUI.ShowUIChrome:
-      if (state.showChrome) return;
-      return { showChrome: true };
-    case StoreActionUI.ShowUISpinner:
-      if (state.showSpinner) return;
-      return { showSpinner: true };
-    default:
-      return;
-  }
-};
 
 interface State { // CHECK: avoid merging here.
   activeExperiments: number;
