+++ webui/react/src/hooks/useTheme.ts
@@ -1,11 +1,12 @@
 import { useCallback, useEffect, useLayoutEffect, useState } from 'react';
 
-import { StoreActionUI, useStore, useStoreDispatch } from 'contexts/Store';
+import { useStore, useStoreDispatch } from 'contexts/Store';
 import useSettings from 'hooks/useSettings';
 import { DarkLight, globalCssVars, Mode } from 'shared/themes';
 import { RecordKey } from 'shared/types';
 import { camelCaseToKebab } from 'shared/utils/string';
 import themes from 'themes';
+import { StoreActionUI } from '../shared/contexts/UIStore';
 
 import { config, Settings } from './useTheme.settings';
