+++ webui/react/src/pages/InteractiveTask.tsx
@@ -3,10 +3,11 @@ import { Helmet } from 'react-helmet-async';
 import { useParams } from 'react-router-dom';
 
 import TaskBar from 'components/TaskBar';
-import { StoreActionUI, useStore, useStoreDispatch } from 'contexts/Store';
+import { useStore, useStoreDispatch } from 'contexts/Store';
 import { getTask } from 'services/api';
 import { CommandState, CommandType } from 'types';
 import handleError from 'utils/error';
+import { StoreActionUI } from '../shared/contexts/UIStore';
 
 import css from './InteractiveTask.module.scss';
 import TaskLogs from './TaskLogs';
