+++ webui/react/src/pages/NoPermissions.tsx
@@ -1,8 +1,9 @@
 import React, { useEffect } from 'react';
 
 import Page from 'components/Page';
-import { StoreActionUI, useStoreDispatch } from 'contexts/Store';
+import { useStoreDispatch } from 'contexts/Store';
 import Icon from 'shared/components/Icon/Icon';
+import { StoreActionUI } from '../shared/contexts/UIStore';
 
 import css from './NoPermissions.module.scss';
